add few test

+++ MeteorologicalSensor/src/main/java/ru/glebdos/MeteorologicalSensor/controllers/MeasurementController.java
@@ -34,7 +34,7 @@ public class MeasurementController {
     }
 
     @GetMapping("/rainyDaysCount")
-    public long getRainyDaysCount() {
+    public Integer getRainyDaysCount() {
         return measurementService.rainyDaysCount();
     }
